# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface CommitteeSettings {
  id: number;
  committeeName: string;
  /** Decimal amount as string */
  monthlyAmount: string;
}

export interface CommitteeSettingsUpdate {
  /** @minLength 1 */
  committeeName?: string;
  /**
     * Decimal amount as string, e.g. "5000.00"
     * @pattern ^\d+(\.\d{1,2})?$
     */
  monthlyAmount?: string;
}

export interface Member {
  id: number;
  name: string;
  phoneNumber: string;
  dateJoined: string;
  createdAt: string;
}

export interface MemberInput {
  /** @minLength 1 */
  name: string;
  /** @minLength 1 */
  phoneNumber: string;
  dateJoined: string;
}

export interface MemberUpdate {
  /** @minLength 1 */
  name?: string;
  /** @minLength 1 */
  phoneNumber?: string;
  dateJoined?: string;
}

export interface Month {
  id: number;
  /** Human readable label, e.g. "January 2026" */
  label: string;
  /** Sortable key, e.g. "2026-01" */
  sortKey: string;
  /** @nullable */
  potWinnerMemberId?: number | null;
  createdAt: string;
}

export interface MonthInput {
  /** @minLength 1 */
  label: string;
  /**
     * Sortable key in YYYY-MM form, e.g. "2026-01"
     * @pattern ^\d{4}-\d{2}$
     */
  sortKey: string;
}

export interface MonthUpdate {
  /** @minLength 1 */
  label?: string;
  /** @nullable */
  potWinnerMemberId?: number | null;
}

export interface Payment {
  id: number;
  memberId: number;
  monthId: number;
  isPaid: boolean;
  /** Decimal amount as string */
  amountPaid: string;
  /** @nullable */
  datePaid?: string | null;
  createdAt: string;
}

export interface PaymentInput {
  memberId: number;
  monthId: number;
  isPaid: boolean;
  /**
     * Decimal amount as string, e.g. "5000.00"
     * @pattern ^\d+(\.\d{1,2})?$
     */
  amountPaid: string;
  /** @nullable */
  datePaid?: string | null;
}

export interface DashboardSummary {
  /** @nullable */
  currentMonthLabel: string | null;
  totalCollectedThisMonth: string;
  totalPendingAmountThisMonth: string;
  pendingMemberCount: number;
  pendingMemberNames: string[];
  totalMembers: number;
  totalMonthsTracked: number;
}