# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: settings
    description: Committee settings
  - name: members
    description: Member management
  - name: months
    description: Month tracking
  - name: payments
    description: Payment tracking
  - name: dashboard
    description: Dashboard summaries
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /settings:
    get:
      operationId: getSettings
      tags: [settings]
      summary: Get committee settings
      responses:
        "200":
          description: Settings
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CommitteeSettings"
    patch:
      operationId: updateSettings
      tags: [settings]
      summary: Update committee settings
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CommitteeSettingsUpdate"
      responses:
        "200":
          description: Updated settings
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CommitteeSettings"

  /members:
    get:
      operationId: listMembers
      tags: [members]
      summary: List all members
      responses:
        "200":
          description: List of members
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Member"
    post:
      operationId: createMember
      tags: [members]
      summary: Create a new member
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MemberInput"
      responses:
        "201":
          description: Created member
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Member"

  /members/{id}:
    get:
      operationId: getMember
      tags: [members]
      summary: Get a member
      parameters:
        - $ref: "#/components/parameters/MemberId"
      responses:
        "200":
          description: Member
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Member"
        "404":
          description: Not found
    patch:
      operationId: updateMember
      tags: [members]
      summary: Update a member
      parameters:
        - $ref: "#/components/parameters/MemberId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MemberUpdate"
      responses:
        "200":
          description: Updated member
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Member"
        "404":
          description: Not found
    delete:
      operationId: deleteMember
      tags: [members]
      summary: Delete a member
      parameters:
        - $ref: "#/components/parameters/MemberId"
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found

  /months:
    get:
      operationId: listMonths
      tags: [months]
      summary: List all tracked months
      responses:
        "200":
          description: List of months
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Month"
    post:
      operationId: createMonth
      tags: [months]
      summary: Add a new month to track (e.g. "Add This Month")
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MonthInput"
      responses:
        "201":
          description: Created month
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Month"

  /months/{id}:
    patch:
      operationId: updateMonth
      tags: [months]
      summary: Update a month (e.g. set pot winner)
      parameters:
        - $ref: "#/components/parameters/MonthId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MonthUpdate"
      responses:
        "200":
          description: Updated month
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Month"
        "404":
          description: Not found
    delete:
      operationId: deleteMonth
      tags: [months]
      summary: Delete a month
      parameters:
        - $ref: "#/components/parameters/MonthId"
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found

  /payments:
    get:
      operationId: listPayments
      tags: [payments]
      summary: List all payments (grid data)
      responses:
        "200":
          description: List of payments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Payment"

  /payments/upsert:
    post:
      operationId: upsertPayment
      tags: [payments]
      summary: Mark payment status for a member/month cell (creates or updates)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PaymentInput"
      responses:
        "200":
          description: Upserted payment
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Payment"

  /dashboard/summary:
    get:
      operationId: getDashboardSummary
      tags: [dashboard]
      summary: Dashboard summary for the current month
      responses:
        "200":
          description: Dashboard summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DashboardSummary"

components:
  parameters:
    MemberId:
      name: id
      in: path
      required: true
      schema:
        type: integer
    MonthId:
      name: id
      in: path
      required: true
      schema:
        type: integer

  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    CommitteeSettings:
      type: object
      required: [id, committeeName, monthlyAmount]
      properties:
        id:
          type: integer
        committeeName:
          type: string
        monthlyAmount:
          type: string
          description: Decimal amount as string

    CommitteeSettingsUpdate:
      type: object
      properties:
        committeeName:
          type: string
          minLength: 1
        monthlyAmount:
          type: string
          pattern: '^\d+(\.\d{1,2})?$'
          description: Decimal amount as string, e.g. "5000.00"

    Member:
      type: object
      required: [id, name, phoneNumber, dateJoined, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        phoneNumber:
          type: string
        dateJoined:
          type: string
          format: date
        createdAt:
          type: string
          format: date-time

    MemberInput:
      type: object
      required: [name, phoneNumber, dateJoined]
      properties:
        name:
          type: string
          minLength: 1
        phoneNumber:
          type: string
          minLength: 1
        dateJoined:
          type: string
          format: date

    MemberUpdate:
      type: object
      properties:
        name:
          type: string
          minLength: 1
        phoneNumber:
          type: string
          minLength: 1
        dateJoined:
          type: string
          format: date

    Month:
      type: object
      required: [id, label, sortKey, createdAt]
      properties:
        id:
          type: integer
        label:
          type: string
          description: Human readable label, e.g. "January 2026"
        sortKey:
          type: string
          description: Sortable key, e.g. "2026-01"
        potWinnerMemberId:
          type: ["integer", "null"]
        createdAt:
          type: string
          format: date-time

    MonthInput:
      type: object
      required: [label, sortKey]
      properties:
        label:
          type: string
          minLength: 1
        sortKey:
          type: string
          pattern: '^\d{4}-\d{2}$'
          description: Sortable key in YYYY-MM form, e.g. "2026-01"

    MonthUpdate:
      type: object
      properties:
        label:
          type: string
          minLength: 1
        potWinnerMemberId:
          type: ["integer", "null"]

    Payment:
      type: object
      required: [id, memberId, monthId, isPaid, amountPaid, createdAt]
      properties:
        id:
          type: integer
        memberId:
          type: integer
        monthId:
          type: integer
        isPaid:
          type: boolean
        amountPaid:
          type: string
          description: Decimal amount as string
        datePaid:
          type: ["string", "null"]
          format: date
        createdAt:
          type: string
          format: date-time

    PaymentInput:
      type: object
      required: [memberId, monthId, isPaid, amountPaid]
      properties:
        memberId:
          type: integer
        monthId:
          type: integer
        isPaid:
          type: boolean
        amountPaid:
          type: string
          pattern: '^\d+(\.\d{1,2})?$'
          description: Decimal amount as string, e.g. "5000.00"
        datePaid:
          type: ["string", "null"]
          format: date

    DashboardSummary:
      type: object
      required:
        [
          currentMonthLabel,
          totalCollectedThisMonth,
          totalPendingAmountThisMonth,
          pendingMemberCount,
          pendingMemberNames,
          totalMembers,
          totalMonthsTracked,
        ]
      properties:
        currentMonthLabel:
          type: ["string", "null"]
        totalCollectedThisMonth:
          type: string
        totalPendingAmountThisMonth:
          type: string
        pendingMemberCount:
          type: integer
        pendingMemberNames:
          type: array
          items:
            type: string
        totalMembers:
          type: integer
        totalMonthsTracked:
          type: integer