# lib/api-zod/src/generated/types/committeeSettingsUpdate.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface CommitteeSettingsUpdate {
  /** @minLength 1 */
  committeeName?: string;
  /**
     * Decimal amount as string, e.g. "5000.00"
     * @pattern ^\d+(\.\d{1,2})?$
     */
  monthlyAmount?: string;
}