# lib/api-zod/src/generated/types/dashboardSummary.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface DashboardSummary {
  /** @nullable */
  currentMonthLabel: string | null;
  totalCollectedThisMonth: string;
  totalPendingAmountThisMonth: string;
  pendingMemberCount: number;
  pendingMemberNames: string[];
  totalMembers: number;
  totalMonthsTracked: number;
}