# lib/api-zod/src/generated/types/memberUpdate.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface MemberUpdate {
  /** @minLength 1 */
  name?: string;
  /** @minLength 1 */
  phoneNumber?: string;
  dateJoined?: Date;
}