# lib/api-zod/src/generated/types/month.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Month {
  id: number;
  /** Human readable label, e.g. "January 2026" */
  label: string;
  /** Sortable key, e.g. "2026-01" */
  sortKey: string;
  /** @nullable */
  potWinnerMemberId?: number | null;
  createdAt: Date;
}