# lib/api-zod/src/generated/types/monthInput.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface MonthInput {
  /** @minLength 1 */
  label: string;
  /**
     * Sortable key in YYYY-MM form, e.g. "2026-01"
     * @pattern ^\d{4}-\d{2}$
     */
  sortKey: string;
}