# lib/api-zod/src/generated/types/monthUpdate.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface MonthUpdate {
  /** @minLength 1 */
  label?: string;
  /** @nullable */
  potWinnerMemberId?: number | null;
}