# lib/api-zod/src/generated/types/payment.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Payment {
  id: number;
  memberId: number;
  monthId: number;
  isPaid: boolean;
  /** Decimal amount as string */
  amountPaid: string;
  /** @nullable */
  datePaid?: Date | null;
  createdAt: Date;
}